rename testSetup to testHelper and move into tests
