+++ backend/tests/integration/genres.test.js
@@ -1,4 +1,4 @@
-const test = require('../../testSetup');
+const test = require('../testHelper');
 const app = require('../../startup/app');
 const { User } = require('../../models/user');
 const { Genre } = require('../../models/genre');
