+++ backend/tests/integration/login.test.js
@@ -1,7 +1,7 @@
 const config = require('config');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
-const test = require('../../testSetup');
+const test = require('../testHelper');
 const app = require('../../startup/app');
 const { User } = require('../../models/user');
