+++ backend/tests/integration/rentals.test.js
@@ -1,5 +1,5 @@
 const mongoose = require('mongoose');
-const test = require('../../testSetup');
+const test = require('../testHelper');
 const app = require('../../startup/app');
 const { User } = require('../../models/user');
 const { Customer } = require('../../models/customer');
