+++ backend/tests/integration/users.test.js
@@ -1,6 +1,6 @@
 const request = require('supertest');
 const bcrypt = require('bcryptjs');
-const test = require('../../testSetup');
+const test = require('../testHelper');
 const app = require('../../startup/app');
 const { User } = require('../../models/user');
